Refactor Calculate Score Button
Refactor calculate score button so it uses the totalScore variable to keep track of the total score value and then return it on the web page. This makes the code more maintainable by only having to use totalScoreEl selector and innerHTML in 2 places for the calculate score button instead of many more places.