Fix Issue with Calculate Score Buttons
Fix an issue where if only zeroes were inputed as answers when you clicked the calculate score button it said Please Answer All Problems instead of telling you your score. This issue was fixed by stoping a few things from being forced to be numbers when they did not need to be.